changed dtat types for dispense and bag vols to double. Worked on dispense logic.

+++ src/com/company/Ice_Cream.java
@@ -2,9 +2,9 @@ package com.company;
 
 public class Ice_Cream {
     public String flavorName;
-    public int bagVol;
+    public double bagVol;
 
-    public Ice_Cream(String flavorName, int bagVol) {
+    public Ice_Cream(String flavorName, double bagVol) {
         this.flavorName = flavorName;
         this.bagVol = bagVol;
     }
